Add files via upload
1. Background of the data and learning task
In this week’s lab session, you will be working on the task of power prediction from an onshore wind 
farm in the north of Germany. The data records the power generated by the wind farm at regular 
intervals, alongside meteorological measurements such as the wind speed at different heights of a 
turbine’s monopile. The original data and website can be found here: 
https://bbdc.csl.uni-bremen.de/index.php/2018h/23-aufgabenstellung-2018