Create Sentiment Analysis using Transfer Learning
This is a transfer learning approach to a sentiment analysis problem. In this project, I did a sentiment analysis of the IMDB dataset using transfer learning and implemented same with some Deep learning approaches and then compared the results. This was done as part of my MSc -AI module submitted to the University of Hull